+++ src/controller/blogs.ts
@@ -1,12 +1,15 @@
 import {sql} from '../db/index';
+const {escape} = require('mysql');
 
 export const getList = async ({author, keyword}: ListRequestParams): Promise<SqlCheckResult> => {
     let sqlText: string = 'select id, title, content, createtime, author from blogs where 1=1';
     if (author) {
-        sqlText += ` and author='${author}'`;
+        author = escape(author);
+        sqlText += ` and author=${author}`;
     }
     if (keyword) {
-        sqlText += ` and title like '%${keyword}%'`;
+        keyword = escape('%' + keyword + '%');
+        sqlText += ` and title like ${keyword}`;
     }
     sqlText += ' order by createtime desc;';
     return await sql(sqlText);
@@ -18,7 +21,8 @@ export const getDetail = async (id: number): Promise<SqlCheckResult> => {
 };
 
 export const updateBlog = async (id: number, content: any): Promise<SqlUpdateResult> => {
-    const sqlText: string = `update blogs set content='${content}' where id=${id};`;
+    content = escape(content);
+    const sqlText: string = `update blogs set content=${content} where id=${id};`;
     return await sql(sqlText);
 };
 
@@ -33,6 +37,10 @@ export const createBlog = async ({
     author,
     createTime
 }: Blog): Promise<SqlInsertResult> => {
-    const sqlText: string = `insert into blogs(title, content, author, createTime, state) values('${title}', '${content}', '${author}', ${createTime}, 1);`;
+    author = escape(author);
+    title = escape(title);
+    content = escape(content);
+    createTime = escape(createTime);
+    const sqlText: string = `insert into blogs(title, content, author, createTime, state) values(${title}, ${content}, ${author}, ${createTime}, 1);`;
     return await sql(sqlText);
 };